% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2tri.R
\name{df2tri}
\alias{df2tri}
\title{df2tri}
\usage{
df2tri(df, node_columns, edge_columns)
}
\arguments{
\item{df}{the dataframe}

\item{node_columns}{node-columns as vector}

\item{edge_columns}{edge-columns as vector. First edge-column should equal edge-weight}
}
\value{
returns edgelist
}
\description{
transforms (wide) `dataframe` into (long) edgelists to work with `igraph`
}
\details{
code based on function `df2nmode` from mardyR package
}
\examples{
\dontrun{
df <- data.frame(from = 1:4,
                 to = c("a", "b", "b", "a"),
                 frame = paste0("f", 4:1),
                 time = paste0("t", c(1,1,2,2)),
                 weight = 1,
                 stringsAsFactors = FALSE)

df2tri(df, node_columns = 1:4, edge_columns = 5)
}
}
