% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_query.R
\name{make_query}
\alias{make_query}
\title{make_query}
\usage{
make_query(keywords)
}
\arguments{
\item{keywords}{vector of keywords in questions}
}
\value{
returns string of query
}
\description{
chains several namespaces or layers into xpath-query
}
\examples{
\dontrun{
make_query(c("custom", "Sentence"))
}
}
