% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_ns.R
\name{select_ns}
\alias{select_ns}
\title{select_ns}
\usage{
select_ns(xmi_file)
}
\arguments{
\item{xmi_file}{path to xmi-file}
}
\value{
returns vector with names
}
\description{
simple wrapper function to extract and print all names of namespaces from xmi-file
}
\examples{
\dontrun{
xmi_file <- unzip_export()
select_ns(xmi_file)
}
}
