% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip_export.R
\name{unzip_export}
\alias{unzip_export}
\title{unzip_export}
\usage{
unzip_export(path, folder, overwrite = FALSE, recursive = FALSE)
}
\arguments{
\item{path}{path of the folder that contains the Inception export}

\item{folder}{name of the folder withing the Inception export, in which the xmi files with the annotations are stored}

\item{overwrite}{should existing unzipped files be unzipped again (defaults to FALSE)}

\item{recursive}{should files from sub-folders also be unzipped (defaults to FALSE)}
}
\value{
unzips files locally in the subfolder "extracted" and provides list with paths to extracted files
}
\description{
unzips files exported from inception (creates local files; use with caution)
}
\examples{
\dontrun{
unzip_export(path = "NameOfExportFolder", folder = "annotation")
}
}
