% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmi2df.R
\name{xmi2df}
\alias{xmi2df}
\title{xmi2df}
\usage{
xmi2df(xmi_file, key = "custom")
}
\arguments{
\item{xmi_file}{xmi-file as provided by inception export}

\item{key}{vector of keywords in questions (default set to 'custom' annotations)}
}
\value{
returns dataframe with queried annotations
}
\description{
extract annotations from inception export (provided as xmi)
}
\examples{
\dontrun{
xmi_files <- list.files(".", pattern = "\\\\.xmi$", recursive = T)
df <- xmi2df(xmi_files[1])
df <- xmi2df(xmi_files[1], c("custom", "Sentence"))
df <- purrr::map_df(xmi_files[1:3], xmi2df, key = c("custom", "Sentence"), .id = "file")
}
}
