% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmi2list.R
\name{xmi2list}
\alias{xmi2list}
\title{xmi2list}
\usage{
xmi2list(xmi_file, key = "custom")
}
\arguments{
\item{xmi_file}{xmi-file as provided by inception export}

\item{key}{vector of keywords in questions (default set to 'custom' annotations)}
}
\value{
returns dataframe with queried annotations
}
\description{
extract text and possibly annotations from inception export (provided as xmi)
}
\examples{
\dontrun{
xmi_files <- list.files(".", pattern = "\\\\.xmi$", recursive = T)
df <- purrr::map_df(xmi_files, purrr::possibly(xmi2list, data.frame(text = NULL)), .id = "id")
}
}
